Primeiro teste com JUnit

+++ TDD e Java - Testes automatizados com JUnit/Testando JUnit/src/br/com/scandura/Calculadora.java
@@ -2,7 +2,7 @@ package br.com.scandura;
 
 public class Calculadora {
 
-    public double somar(double a, double b){
+    public int somar(int a, int b){
         return a + b;
     }
